Added scroll button and updated website description
Added a scroll button that appears after the page is scrolled 20px from the top. When it is clicked, it jumps to the top of the page. The CSS scroll behavior property was set as smooth.

Changed JS to use querySelector in all cases instead of getElementById.

Re-wrote website description and added a meta description.

Removed "close-nav" class from HTML. The navigation button still seems to work.