Changed navigation and color scheme
HTML - Added menu that includes navigation button and scroll-to-top button.  Moved navigation out of header tag, so that it is a separate entity.
CSS - Added root to create CSS variables for colors.  Added overlay class to show an overlay over body when navigation is open.  Changed color scheme from bright blue to a darker blue.
JS - Added overlay class and change nav button from bars to times when <nav> is open.  Close nav and remove overlay when overlay is clicked or page is scrolled to top.